+++ src/client.ts
@@ -5,15 +5,15 @@
  * Licensed under the LGPLv3 License. See LICENSE.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 'use strict';
-import {
-  LanguageClient
-} from 'vscode-languageclient';
-import {
-  RulesResponse
-} from './rules';
+import { LanguageClient } from 'vscode-languageclient';
+import { RulesResponse } from './rules';
 
 export class SonarLintExtendedLanguageClient extends LanguageClient {
   listAllRules(): Thenable<RulesResponse> {
     return this.sendRequest('sonarlint/listAllRules');
+  }
+
+  didClasspathUpdate(projectRoot: string): void {
+    this.sendNotification('sonarlint/didClasspathUpdate', projectRoot);
   }
 }